Spring Data JPA

+++ learn-database/frame-relational/orm-spring-data-jpa/000-jap-hello-world/src/main/java/com/code/jpa/dal/entity/base/AbstractEntity.java
@@ -22,13 +22,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
+import javax.persistence.*;
 import java.io.Serializable;
 import java.util.Date;
 
@@ -41,7 +35,7 @@ import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
 @MappedSuperclass
-//@EntityListeners(EntityListener.class)
+// @EntityListeners(EntityListener.class)
 public abstract class AbstractEntity implements Serializable {
 
 	/**
@@ -49,7 +43,7 @@ public abstract class AbstractEntity implements Serializable {
 	 */
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 
 	/**
 	 * 创建时间
